Invoice-management-tool/requirements.txt
flask==3.1.1
google-api-python-client==2.166.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
reportlab==4.4.0
openpyxl==3.1.5
